♻️ F3Q401 guard: decompose close messages
