+++ .agentplane/tasks/202605282325-F3Q401/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605282325-F3Q401`
+Title: Close commit message decomposition
+Canonical task record: `.agentplane/tasks/202605282325-F3Q401/README.md`
+
+## Summary
+
+Close commit message decomposition
+
+Decompose packages/agentplane/src/commands/guard/impl/close-message.ts by extracting git metadata and rendering helpers while preserving buildCloseCommitMessage/renderMergeMessage behavior.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/commands/guard/impl/close-message.ts by extracting git metadata and rendering helpers while preserving buildCloseCommitMessage/renderMergeMessage behavior.
+- Out of scope: unrelated refactors not required for "Close commit message decomposition".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-28T23:26:21.787Z
+- Branch: task/202605282325-F3Q401/close-message-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
