+++ .agentplane/tasks/202605282325-F3Q401/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 F3Q401 task: Close commit message decomposition [202605282325-F3Q401]
